* Uploaded some artwork relating to the friends who made this program possible. Will be used in the About screen! :) * Updated the About screen, as per the above. * Coded up some further functionality relating to the new spectrograph / waterfall code and the KissFFT functions. We nearly have a functioning waterfall at this point!

+++ src/gk_fft.hpp
@@ -72,7 +72,7 @@ public:
     explicit GkFFT(QPointer<GekkoFyre::GkEventLogger> eventLogger, QObject *parent = nullptr);
     ~GkFFT();
 
-    std::vector<GekkoFyre::Spectrograph::GkFFTSpectrum> FFTCompute(const std::vector<float> &data, const GekkoFyre::Database::Settings::Audio::GkDevice &audioDevice,
+    std::vector<GekkoFyre::Spectrograph::GkFFTSpectrum> FFTCompute(const std::vector<double> &data, const GekkoFyre::Database::Settings::Audio::GkDevice &audioDevice,
                                                                    const qint32 &numSamples);
     std::vector<float> FFTCurvePlot(const std::vector<float> &data, const GekkoFyre::Database::Settings::Audio::GkDevice &audioDevice,
                                     const qint32 &numSamples);
